Rectify SonarQube Issues
SonarQube identified some small issues within the code. This commit rectifies these issues.
- Remove unnecessary unboxing for financialAssessmentId in EvidenceOrchestrationService
- Replace Collectors.toList() with Stream.toList() in TestModelDataBuilder as this prevents list from being modified.
- Rename variable from ContributionSummaryDTOs to contributionSummaryDTOs in ContributionMapperTest to follow java naming standards.

+++ maat-orchestration/src/main/java/uk/gov/justice/laa/crime/orchestration/service/orchestration/EvidenceOrchestrationService.java
@@ -60,11 +60,8 @@ public class EvidenceOrchestrationService {
         Date upliftRemovedDate =
                 Objects.requireNonNullElse(incomeEvidenceSummaryDTO.getUpliftRemovedDate(), defaultDate);
 
-        Integer financialAssessmentId = applicationDTO
-                .getAssessmentDTO()
-                .getFinancialAssessmentDTO()
-                .getId()
-                .intValue();
+        Integer financialAssessmentId =
+                applicationDTO.getAssessmentDTO().getFinancialAssessmentDTO().getId();
         FinancialAssessmentDTO financialAssessmentDTO = repOrderDTO.getFinancialAssessments().stream()
                 .filter(assessment -> assessment.getId().equals(financialAssessmentId))
                 .findFirst()
